Move NameBoxCalculator to core
